+++ core/src/main/java/io/github/cactacea/core/domain/enums/GroupInvitationStatusType.java
@@ -3,17 +3,17 @@ package io.github.cactacea.core.domain.enums;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public enum GroupInvitationStatusType {
-    @JsonProperty("0") noresponsed(0),
-    @JsonProperty("1") accepted(1),
-    @JsonProperty("2") rejected(2);
+    @JsonProperty("0") noresponsed((byte)0),
+    @JsonProperty("1") accepted((byte)1),
+    @JsonProperty("2") rejected((byte)2);
 
-    private long value;
+    private byte value;
 
-    private GroupInvitationStatusType(long value) {
+    private GroupInvitationStatusType(byte value) {
         this.value = value;
     }
 
-    static public GroupInvitationStatusType forName(long value) {
+    static public GroupInvitationStatusType forName(byte value) {
         for (GroupInvitationStatusType e : values()) {
             if (e.value == value) {
                 return e;
@@ -22,7 +22,7 @@ public enum GroupInvitationStatusType {
         throw new IllegalArgumentException();
     }
 
-    public long toValue() {
+    public byte toValue() {
         return value;
     }
 }
